{knowledge,memory}: split sqlitevec example implementations
Move the sqlitevec implementations out of util.go and into CGO-gated files so downstream synced copies do not collide with existing stub files. Update examples/knowledge/go.mod to depend on a version that includes the extractor packages used by the synced example.
